move quality judge in a subcommand
